edit

+++ model/request/wechat.go
@@ -2,7 +2,7 @@ package request
 
 type CheckSignature struct {
 	Signature string `json:"signature" form:"signature"`
-	Timestamp int16  `json:"timestamp" form:"timestamp"`
+	Timestamp int    `json:"timestamp" form:"timestamp"`
 	Nonce     string `json:"nonce" form:"nonce"`
 	Echostr   string `json:"echostr" form:"echostr"`
 }
